i18n for shared modules
